// src/lib/bindings/GeneratedToolDefinition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { OpenApiParameter } from "./OpenApiParameter";

export type GeneratedToolDefinition = {
  toolName: string;
  label: string;
  description: string;
  method: string;
  path: string;
  parameters: Array<OpenApiParameter>;
};